--- include/modelgenerator.hpp
@@ -1,5 +1,0 @@
-#ifndef GAMEOFLIFE_MODELGENERATOR_H
-#define GAMEOFLIFE_MODELGENERATOR_H
-
-
-#endif //GAMEOFLIFE_MODELGENERATOR_H
