+++ src/ui/interface.hpp
@@ -1,7 +1,7 @@
 #ifndef GAMEOFLIFE_INTERFACE_HPP
 #define GAMEOFLIFE_INTERFACE_HPP
 
-#include <modelpresets.hpp>
+#include <presets\modelpresets.hpp>
 
 #include <functional>
 #include <memory>
